requirements.txt
# Django
Django==3.1.4

# Postgres
psycopg2==2.8.6
dj-database-url==0.5.0

# Channels
channels==3.0.2
channels-redis==3.2.0

# Rest Framework
djangorestframework==3.12.2

# Media files
Pillow==8.0.1

# Auth
djangorestframework-simplejwt==4.6.0